Actualizando factories

+++ database/migrations/2020_11_25_175936_create_clients_table.php
@@ -17,7 +17,7 @@ class CreateClientsTable extends Migration
             $table->id();
             $table->string('code')->unique();
             $table->string('name');
-            $table->integer('city_id')->unsigned()->nullable();
+            $table->bigInteger('city_id')->unsigned()->nullable();
             $table->foreign('city_id')->references('id')->on('cities')->onDelete('cascade');
             $table->timestamps();
         });
